requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --no-emit-options --output-file=requirements.txt requirements.in
#
anyio==3.6.2
    # via starlette
bcrypt==4.0.1
    # via -r requirements.in
click==8.1.3
    # via uvicorn
databases==0.6.2
    # via -r requirements.in
dnspython==2.2.1
    # via pymongo
ecdsa==0.18.0
    # via python-jose
fastapi==0.87.0
    # via -r requirements.in
greenlet==2.0.1
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
motor==3.1.1
    # via -r requirements.in
passlib==1.7.4
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pydantic==1.10.2
    # via fastapi
pyjwt==2.6.0
    # via -r requirements.in
pymongo==4.3.2
    # via motor
python-dotenv==0.21.0
    # via -r requirements.in
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
rsa==4.9
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   python-multipart
sniffio==1.3.0
    # via anyio
sqlalchemy==1.4.41
    # via databases
starlette==0.21.0
    # via fastapi
typing-extensions==4.4.0
    # via pydantic
uvicorn==0.19.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
